one more change

+++ qutip/about.py
@@ -93,39 +93,40 @@ def about():
 
 def tk_conify_center():
     import os
-    try: os.environ['FRANCO']=='TRUE'
-    except: return
+    try: os.environ['FRANCO']
+    except: pass
     else:
-        import Tkinter,zipfile,time
-        def center(window):
-          sw = window.winfo_screenwidth()
-          sh = window.winfo_screenheight()
-          rw = window.winfo_reqwidth()
-          rh = window.winfo_reqheight()
-          xc = (sw - rw) / 2
-          yc = (sh -rh) / 2
-          window.geometry("+%d+%d" % (xc-75, yc-75))
-          window.deiconify()
-        def stop(me):
-            stop_flag=1
-            me.destroy()
-            os.remove(os.path.dirname(__file__)+'/.egg.gif')
-            os.environ['FRANCO']='FALSE'
-        root=Tkinter.Tk() 
-        root.title('The Franco Easter Egg')
-        root.wm_attributes("-topmost", 1)
-        zf=zipfile.ZipFile(os.path.dirname(__file__)+"/.Tk.egg.zip", 'r')
-        data=zf.extract('.egg.gif',os.path.dirname(__file__),pwd='lowfruit')
-        c=Tkinter.Canvas(root,width=290, height=300) 
-        p=Tkinter.PhotoImage(file=data) 
-        i=c.create_image(0,0,anchor=Tkinter.NW,image=p) 
-        c.pack() 
-        root.after(0,center,root)
-        root.after(5000,stop,root)    
-        root.mainloop()
-        try:os.remove(os.path.dirname(__file__)+'/.egg.gif')
-        except:os.environ['FRANCO']='FALSE'
-        else:os.environ['FRANCO']='FALSE'
+        if os.environ['FRANCO']=='TRUE':
+            import Tkinter,zipfile,time
+            def center(window):
+              sw = window.winfo_screenwidth()
+              sh = window.winfo_screenheight()
+              rw = window.winfo_reqwidth()
+              rh = window.winfo_reqheight()
+              xc = (sw - rw) / 2
+              yc = (sh -rh) / 2
+              window.geometry("+%d+%d" % (xc-75, yc-75))
+              window.deiconify()
+            def stop(me):
+                stop_flag=1
+                me.destroy()
+                os.remove(os.path.dirname(__file__)+'/.egg.gif')
+                os.environ['FRANCO']='FALSE'
+            root=Tkinter.Tk() 
+            root.title('The Franco Easter Egg')
+            root.wm_attributes("-topmost", 1)
+            zf=zipfile.ZipFile(os.path.dirname(__file__)+"/.Tk.egg.zip", 'r')
+            data=zf.extract('.egg.gif',os.path.dirname(__file__),pwd='lowfruit')
+            c=Tkinter.Canvas(root,width=290, height=300) 
+            p=Tkinter.PhotoImage(file=data) 
+            i=c.create_image(0,0,anchor=Tkinter.NW,image=p) 
+            c.pack() 
+            root.after(0,center,root)
+            root.after(5000,stop,root)    
+            root.mainloop()
+            try:os.remove(os.path.dirname(__file__)+'/.egg.gif')
+            except:os.environ['FRANCO']='FALSE'
+            else:os.environ['FRANCO']='FALSE'
 
 
 if __name__ == "__main__":
